refactor(affine-transforms): move affine transformation functionality from Model class to own module
