+++ lib/src/main/java/io/github/nonmilk/coffee/grinder/transformations/RotationOrder.java
@@ -1,0 +1,5 @@
+package io.github.nonmilk.coffee.grinder.transformations;
+
+public enum RotationOrder {
+    XYZ, XZY, YXZ, YZX, ZXY, ZYX;
+}
